# integration-framework/moirai-zeus/src/main/resources/static/swagger/apis/openapi.yaml
openapi: 3.0.0
info:
  title: Moirai Zeus API
  description: This is the API for the Zeus(Ta3) system's access to the subprojects (e.g. Clotho),
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
security:
  - bearerAuth: []
paths:
...
  /admin/assumeUser:
    post:
      tags:
        - Admin
      summary: Logs the admin in as the user.
      description: A logged in admin can use this endpoint to assume the login of the desired user.
      operationId: assumeUser
      requestBody:
        description: Desired Username
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/stringRequest'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /admin/activateDeactivateAccounts:
    put:
      tags:
        - Admin
      summary: DeActivates or Activates a Users Account
      description: Deactivates or activates a users account
      operationId: toggleAccountActivation
      parameters:
        - name: userName
          in: query
          description: User to flip Account Active flag
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /admin/listRoles:
    get:
      tags:
        - Admin
      summary: Returns JwtRole list
      description: Returns all Roles stored for displaying and updating.
      operationId: listRoles
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                items:
                  $ref: '#/components/schemas/jwtRole'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /admin/deleteDockerRegistry:
    delete:
      tags:
        - Admin
      summary: Takes a user and updates their teamname across the schema
      description: creates or updates existing users team
      operationId: deleteTeamRegistry
      parameters:
        - name: teamName
          in: query
          description: teamName to assign to user
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        404:
          description: Not Found. Registry was not found.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /admin/updateUsersTeam:
    put:
      tags:
        - Admin
      summary: Takes a user and updates their teamname across the schema
      description: creates or updates existing users team
      operationId: updateTeamName
      parameters:
        - name: teamName
          in: query
          description: teamName to assign to user
          schema:
            type: string
      requestBody:
        description: user to save
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/jwtUser'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        409:
          description: Conflict
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /admin/mergeTeam:
    put:
      tags:
        - Admin
      summary: Migrates teamname across the schema
      description: creates or updates existing users team
      operationId: mergeTeam
      parameters:
        - name: teamName
          in: query
          description: teamName to assign to user
          schema:
            type: string
      requestBody:
        description: user to save
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/jwtUser'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        409:
          description: Conflict
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /admin/createOrUpdateRole:
    post:
      tags:
        - Admin
      summary: Takes a user and updates their Roles.
      description: creates or updates existing Roles
      operationId: createOrUpdateRole
      requestBody:
        description: roles to save
        content:
          'application/json':
            schema:
              items:
                $ref: '#/components/schemas/jwtRole'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        409:
          description: Conflict
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /admin/updateUserAccount:
    put:
      tags:
        - Admin
      summary: Takes a username and Role list.
      description: Assigns a Users account with Roles passed
      operationId: updateUserAccount
      requestBody:
        description: JwtUser to update
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/userDataDto'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /admin/deleteRoles:
    delete:
      tags:
        - Admin
      summary: Delete a Role.
      description: deletes a Role and removes that role from all users assigned to it
      operationId: deleteRoles
      requestBody:
        description: roles to delete
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/jwtRole'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        404:
          description: Role Not Found
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        409:
          description: Conflict, the role may have been modified after you selected it for deletion.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /admin/listUsers:
    get:
      tags:
        - Admin
      summary: Retrieves a list of users and their permissions.
      description: list users
      operationId: listUsers
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                items:
                  $ref: '#/components/schemas/jwtUser'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /admin/listPermissions:
    get:
      tags:
        - Admin
      summary: Retrieves a list of permissions.
      description: list permissions
      operationId: listPermissions
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                items:
                  $ref: '#/components/schemas/jwtPermission'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /admin/provisionGitLab:
    post:
      tags:
        - Admin
      summary: Provisions Gitlab and Gitlab runner for project needs
      description: Provisions Gitlab and Gitlab runner for project needs
      operationId: provisionGitlab
      requestBody:
        description: Launches Terraform with the selected parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/gitLabRequest'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
    delete:
      tags:
        - Admin
      summary: Destroys GitLab and Gitlab runner for project needs
      description: Destroys Gitlab and Gitlab runner for project needs
      operationId: destroyGitlab
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /admin/provisionEnclaveInfrastructure:
    post:
      tags:
        - Admin
      summary: Provisions Enclace Infrastructure for project needs
      description: Provisions Enclace Infrastructure for project needs
      operationId: provisionEnclaveInfrastructure
      requestBody:
        description: Launches Terraform with the selected parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/stringRequest'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
    delete:
      tags:
        - Admin
      summary: Destroys Enclave Infrastructure for project teardown
      description: Destroys Enclave Infrastructure for project teardown
      operationId: destroyEnclaveInfrastructure
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    gitLabRequest:
      type: object
      required:
        - prop1
      properties:
        prop1:
          type: string
        prop2:
          type: string
...
    userDataDto:
      type: object
      required:
        - id
        - username
        - performerGroup
        - emailAddress
        - active
      properties:
        id:
          type: string
        username:
          type: string
          example: "persongroup"
        performerGroup:
          type: string
          example: "TA1"
        emailAddress:
          type: string
          example: "generic.email@host.com"
        active:
          type: boolean
          example: true
        roles:
          type: array
          items:
            $ref: '#/components/schemas/jwtRole'
    jwtUser:
      x-jpaEntity: true
      type: object
      required:
        - username
        - password
        - emailAddress
        - teamName
        - performerGroup
      properties:
        id:
          type: string
          example:
          x-jpaIdField: true
          x-generatedId: true
        active:
          type: boolean
          x-jpaColOverride:
            overrideValue: "columnDefinition = \"tinyint(1) default 1\""
        username:
          type: string
          example: "moirai_group"
        password:
          type: string
          example: "AVerySecurePassword"
        passwordExpiration:
          type: string
          format: date
          example: "2020-01-01"
        teamName:
          type: string
          example: "Next Century"
        performerGroup:
          type: string
          example: "TA1"
        emailAddress:
          type: string
          example: "generic.email@host.com"
        profilePictureURL:
          type: string
        darkMode:
          type: boolean
          example: true
        roles:
          type: array
          items:
            $ref: '#/components/schemas/jwtRole'
          x-lazyManyToMany: true
          x-joinTable:
            joinValue: name = "jwt_user_jwt_role",
              joinColumns = { @JoinColumn(name = "jwt_user_id") },
              inverseJoinColumns = { @JoinColumn(name = "jwt_role_id") }
        timezone:
          type: string
          example: "GMT-05:00"
        services:
          type: array
          items:
            $ref: '#/components/schemas/service'
          x-lazyOneToMany: true
          x-jsonIgnoreProperty:
            propertyKey: missions
          x-cascade-all: true
    stringResponse:
      type: object
      properties:
        value:
          type: string
          example: "Success"
    stringRequest:
      type: object
      properties:
        value:
          type: string
          example: "Sample_text"
    jwtPermission:
      type: object
      x-jpaEntity: true
      required:
        - id
        - name
        - description
      properties:
        id:
          type: string
          example:
          x-jpaIdField: true
          x-generatedId: true
        name:
          type: string
          example: "can_provision_graph_db"
        description:
          type: string
          example: "Has access to the provisioning service for graph databases."
    jwtRole:
      type: object
      x-jpaEntity: true
      required:
        - id
        - name
        - description
      properties:
        id:
          type: string
          example:
          x-jpaIdField: true
          x-generatedId: true
        name:
          type: string
          example: "admin"
        description:
          type: string
          example: "Administrative User"
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/jwtPermission'
          x-lazyManyToMany: true
          x-jsonIgnoreProperty:
            propertyKey: missions
          x-joinTable:
            joinValue: name = "jwt_permission_jwt_role",
              joinColumns = { @JoinColumn(name = "jwt_role_id") },
              inverseJoinColumns = { @JoinColumn(name = "jwt_permission_id") }
    service:
      type: object
      x-jpaEntity: true
      required:
        - name
      properties:
        id:
          type: string
          example:
          x-jpaIdField: true
          x-generatedId: true
        name:
          type: string
          example: "service1"
        type:
          type: string
          example: "clotho"
        subtype:
          type: string
          example: "Neo4j"
        status:
          type: string
        dateReceived:
          type: string
          format: date
          example: "2020-01-01"
        teamName:
          type: string
        access:
          type: string
          example: "Private"
        details:
          type: string
        error:
          type: string
          example: "Error Msg"
          x-jpaTransient: true
        awsInstances:
          type: array
          items:
            $ref: '#/components/schemas/serviceAwsInstance'
          x-lazyOneToMany: true
          x-jsonIgnoreProperty:
            propertyKey: missions
          x-joinTable:
            joinValue: name = "service_service_aws_instance",
              joinColumns = { @JoinColumn(name = "service_id") },
              inverseJoinColumns = { @JoinColumn(name = "service_aws_instance_id") }
          x-cascade-all: true
        endpoints:
          type: array
          items:
            $ref: '#/components/schemas/serviceEndpoint'
          x-lazyOneToMany: true
          x-jsonIgnoreProperty:
            propertyKey: missions
          x-joinTable:
            joinValue: name = "service_service_endpoint",
              joinColumns = { @JoinColumn(name = "service_id") },
              inverseJoinColumns = { @JoinColumn(name = "service_endpoint_id") }
          x-cascade-all: true
        downloads:
          type: array
          items:
            $ref: '#/components/schemas/serviceDownload'
          x-lazyOneToMany: true
          x-jsonIgnoreProperty:
            propertyKey: missions
          x-joinTable:
            joinValue: name = "service_service_download",
              joinColumns = { @JoinColumn(name = "service_id") },
              inverseJoinColumns = { @JoinColumn(name = "service_download_id") }
          x-cascade-all: true
    serviceAwsInstance:
      type: object
      x-jpaEntity: true
      required:
        - instanceId
      properties:
        id:
          type: string
          x-jpaIdField: true
          x-generatedId: true
        instanceId:
          type: string
          example: "i-011d626d652684d5a4"
    serviceEndpoint:
      type: object
      x-jpaEntity: true
      required:
        - uri
        - name
      properties:
        id:
          type: string
          x-jpaIdField: true
          x-generatedId: true
        name:
          type: string
          example: "Clotho Endpoint"
        uri:
          type: string
          example: "ec2-18-232-154-122.compute-1.amazonaws.com:8008"
    serviceDownload:
      type: object
      x-jpaEntity: true
      required:
        - uri
        - name
      properties:
        id:
          type: string
          x-jpaIdField: true
          x-generatedId: true
        name:
          type: string
          example: "Python Connector"
        uri:
          type: string
          example: "s3://kairos-service-artifacts/connectors/next_century/service_a/connector.py"
...